+++ src/main/java/com/decagon/queuepay/repositories/UserRepository.java
@@ -4,11 +4,10 @@ import com.decagon.queuepay.models.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
-import java.util.UUID;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByUsername(String username);
-    Boolean existsByUsername(String username);
-    User findByEmailVerificationToken(String token);
+public interface UserRepository extends JpaRepository<User, Long> {
+  Optional<User> findByUsername(String username);
+  Boolean existsByUsername(String username);
+  User findByEmailVerificationToken(String token);
 }
